+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
--
+### Fixed
+- version check when offline
 
 ## [0.3.0] - 2016-10-15
 ### Added